requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
python-multipart
requests
SpeechRecognition
gTTS
pydantic
pydub 